+++ yash-semantics/CHANGELOG.md
@@ -7,10 +7,38 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.8.1] - Unreleased
 
+### Added
+
+- The `command_search::classify` function has been added to classify a command
+  name without searching for an external utility.
+    - This function is used in the execution of a simple command
+      (`impl command::Command for yash_syntax::syntax::SimpleCommand`) to classify
+      the command name before performing redirections and variable assignments
+      without redundant searching for an external utility.
+- The `command_search::ClassifyEnv` trait has been added to provide the
+  environment required by the `command_search::classify` function.
+    - This trait is a subset of the `command_search::SearchEnv` trait.
+- Internal dependencies:
+    - either 1.9.0
+
 ### Changed
 
 - External dependency versions:
     - yash-syntax 0.15.0 → 0.15.1
+
+### Removed
+
+- Internal dependencies:
+    - assert_matches 1.5.0
+
+### Fixed
+
+- The execution of a simple command
+  (`impl command::Command for yash_syntax::syntax::SimpleCommand`)
+  now searches for the external utility in the `PATH` after performing variable
+  assignments, as specified in POSIX.1-2024. Previously, it would search for the
+  utility before performing the redirections and assignments, which could lead
+  to incorrect behavior if the assignments modified the `PATH` variable.
 
 ## [0.8.0] - 2025-05-11